+++ warrior.py
